refactor(initialization): extract the handler initialisation steps
guard_core/core/initialization/handler_initializer.py was rank B (MI 18.71).
The Redis-connect, cloud/geo-without-redis, lazy-init and agent-wiring
steps of HandlerInitializer (initialize_redis_handlers,
initialize_agent_integrations, initialize_agent_for_handlers,
initialize_dynamic_rule_manager, shutdown_agent_integrations, and their
private helpers) move to a new _HandlerInitializerStepsMixin in
_handler_initializer_steps.py; HandlerInitializer inherits it and keeps
the builder methods (build_enricher, build_composite_handler,
build_event_filter, build_event_bus, build_metrics_collector) and
get_initialization_status. HandlerInitializer keeps every method it had
before; nothing importable moved.

Sync mirror regenerated.
